requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.128.0
filelock==3.20.1
fsspec==2025.12.0
h11==0.16.0
idna==3.11
jinja2==3.1.6
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
pillow==12.0.0
pydantic==2.12.5
pydantic-core==2.41.5
starlette==0.50.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
